Update sensitiveness of the Apply button in 'download_driver' mode
At the moment, the Forward button shown in the 'download drivers' page
is properly updating its sensitiveness depending on whether the license
has been explicitly accepted or not. However, the Apply button shown
when running the NewPrinterGUI dialog in 'download_driver' mode is
always sensitive, regardless of the license being accepted.

This patch fixes this issue by sharing the code that checks the
acceptance of the license between the two use cases.

https://fedorahosted.org/system-config-printer/ticket/238