Use eosdriverinstaller.py to automatically install printer driver
We will use the DriverInstallerThread class to make sure we install the
drivers from OpenPrinting in a separate thread, so that we can keep
updating the UI in the meanwhile manually.

We do it this way instad of using the usual asynchronous way so that we
can more easily integrate with the upstream code from s-c-p, which
relies in a different script (install-printerdriver.py) we don't need.

[endlessm/eos-shell#3971]