Always query openprinting, regardless of the local drivers available
In some cases we do have a suitable driver pre-installed that gets
categorized as an 'exact' fit for the printer being configured, meaning
that system-config-printer will simply pick it up without looking for
alternative options in OpenPrinting.org, regardless of some better option
being available from there.

For instance, this happens with the Epson XP-412 printer for which we
already ship a suitable driver pre-installed (ESC/P-R), which works but is
not as good as the one recommended by OpenPrinting.org, the ESC/P driver.

This patch makes sure that s-c-p always queries OpenPrinting.org (should an
Internet connection is available) for drivers for the printer being installed,
so that it makes sure to give users an option to download them before going
ahead with the pre-installed one.

If an Internet connection is not available or the user refuses/cancels
downloading the recommended driver, the pre-installed one will be used.

https://phabricator.endlessm.com/T10886