Make sure we give recommended printer drivers priority when installing
Sometimes it might happen that you have more than one driver available
for the same printer, both of them categorized as an 'exact' fit for the
printer, in which case it's not deterministic which one s-c-p will end
up selecting when installing the printer.

For instance this might happen when installing an Epson printer that
is both covered by the generic ESC/P-R driver and the vendor-specific
ESC/P one, either pre-installed or downloaded from OpenPrinting.org.

To help s-c-p make a decision in this case, we define additional matching
rules to detect this kind of drivers (Canon, Epson and OpenPrinting ones
for now) so that they are given priority against drivers for the same
printer with the same 'fitting category', with the only exception of
those categorized as 'exact-cmd' which are the strongest possible match.

Note: There's a chance that this fixes the issue related to the Canon
PIXMA MG2410 printer, since it will make sure that the pre-installed
vendor-specific driver gets selected, instead of the Gutenprint one.

https://phabricator.endlessm.com/T6295
https://phabricator.endlessm.com/T10886