Give drivers installed via eos-config-printer priority when installing
We need a way to decide which driver to install for those cases that we
have the same option available both via pre-installed drivers and via those
downloaded via eos-config-printer, which in some cases might collide.

Considering that drivers downloaded using eos-config-printer will never be
older than the ones we pre-install, we give them priority at the time of
configuring the printer, so that we always use the most recent one.

https://phabricator.endlessm.com/T10886