Do not query OpenPrinting.org for drivers that have already been installed
This is a quick sanity quick check so that we don't query remote sources
forever for drivers that we either downloaded and installed already (using
eos-config-printer) or for drivers that, while being available via remote
sources, we have explicitly chosen to pre-install them in our images.

Ideally, we should have a declarative way to do this instead of hard coding
it, but the truth is that this will only affect binary drivers containing
'OpenPrinting' in the PPD name, or those LSB-compliant drivers for Epson and
Canon Inkjet printers that we ship (see http://www.openprinting.org/drivers),
as well as those automatically downloaded and installed via eos-config-printer,
so this should be enough while keeping our local changes as minimal as possible.

https://phabricator.endlessm.com/T10886