ppds: skip self-test if CUPS is not running
The Endless build server does not run the CUPS daemon, so this test
would always fail with:

    Traceback (most recent call last):
      File "/build/src/system-config-printer_1.5.5+dev81.21f9f3b/_build/sub/cupshelpers/ppds.py", line 1247, in _self_test
        with open (picklefile, "rb") as f:
    FileNotFoundError: [Errno 2] No such file or directory: 'pickled-ppds'

    During handling of the above exception, another exception occurred:

    Traceback (most recent call last):
      File "./test-ppd-module.sh", line 5, in <module>
        _self_test(sys.argv)
      File "/build/src/system-config-printer_1.5.5+dev81.21f9f3b/_build/sub/cupshelpers/ppds.py", line 1251, in _self_test
        c = cups.Connection ()
    RuntimeError: failed to connect to server

Work around this by aborting the test early if connecting to CUPS fails.
Autotools understands exit code 77 to mean "skipped".
https://www.gnu.org/software/automake/manual/automake.html#Scripts_002dbased-Testsuites-1

Do not create the picklefile in this case – otherwise, the next test run
to try to load it and fail with a different, unhandled exception.

Not submitting upstream because upstream commit
9b2fdc04c0c2e8cc1acae47e238fdb2f3bef7f53 reworks this entire test in an
analogous but more invasive way.